txengine: move code to new module
