+++ electrum/lnwatcher.py
@@ -279,4 +279,4 @@ class LNWalletWatcher(LNWatcher):
             # careful, this prevents revocation as well
             if not self.lnworker.enable_htlc_settle_onchain:
                 return
-        self.lnworker.wallet.add_sweep_info(sweep_info)
+        self.lnworker.wallet.txengine.add_sweep_info(sweep_info)
